RLC: Logical Operators. Compilation of Stack Map Frames

+++ rlc/app/src/nativeMain/kotlin/org/example/rlc/jvm/ir/FullFrameBuilder.kt
@@ -1,12 +1,12 @@
 package org.example.rlc.jvm.ir
 
 internal class FullFrameBuilder {
-  private var offsetDelta: Byte = 0
+  private var offsetDelta: Short = 0
   private var locals: List<VerificationTypeInfo> = listOf()
   private var stack: List<VerificationTypeInfo> = listOf()
   private var needToBuild = false
 
-  fun offsetDelta(value: Byte): FullFrameBuilder {
+  fun offsetDelta(value: Short): FullFrameBuilder {
     offsetDelta = value
     return this
   }
